+++ tests/unit/metadata/DockerComposeMetadata.spec.js
@@ -1,21 +1,16 @@
-import DockerComposeMetadata from 'src/metadata/DockerComposeMetadata';
-import { DefaultData } from 'leto-modelizer-plugin-core';
+import { getComposatorMetadata } from 'tests/resources/metadata/getComposatorMetadata';
 
-describe('Test class: DockerComposeMetadata', () => {
-  describe('Test method: validate', () => {
-    it('should return true', () => {
-      expect(new DockerComposeMetadata().validate()).toEqual(true);
-    });
-  });
-
-  describe('Test method: parse', () => {
-    it('should set components definitions to empty array', () => {
-      const pluginData = new DefaultData();
-      pluginData.definitions.components = ['a'];
-
-      new DockerComposeMetadata(pluginData).parse();
-
-      expect(pluginData.definitions.components).toEqual([]);
+describe('Test DockerComposeMetadata', () => {
+  describe('Test methods', () => {
+    describe('Test method: validate', () => {
+      it('Should return true on valid metadata', () => {
+        const metadata = getComposatorMetadata('validMetadata', 'tests/resources/metadata/valid.json');
+        expect(metadata.validate()).toBeTruthy();
+      });
+      it('Should return false on invalid metadata', () => {
+        const metadata = getComposatorMetadata('invalidMetadata', 'tests/resources/metadata/invalid.json');
+        expect(metadata.validate()).toBeFalsy();
+      });
     });
   });
 });