+++ changelog.md
@@ -5,8 +5,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html)
 
-## [Unreleased]
+## [0.1.0]
 
 ### Added
 
-- Setup project.
+- Add yaml syntax highlighting
+- Support for parsing Docker-Compose files.
+- Add Docker-Compose metadata.
+- leto-modelizer-plugin-core version 0.21.0